Fix compile issues

+++ src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/Vector128.cs
@@ -2413,7 +2413,7 @@ namespace System.Runtime.Intrinsics
                 {
                     selectedValue = vector.GetElementUnsafe(selectedIndex);
                 }
-                result.SetElementUnsafe(index, selectedValue);
+                result.SetElementUnsafe(index, (sbyte)selectedValue);
             }
 
             return result;
