+++ src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/Vector512.cs
@@ -2437,7 +2437,7 @@ namespace System.Runtime.Intrinsics
                 {
                     selectedValue = vector.GetElementUnsafe(selectedIndex);
                 }
-                result.SetElementUnsafe(index, selectedValue);
+                result.SetElementUnsafe(index, (sbyte)selectedValue);
             }
 
             return result;
